pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "veritydocs"
version = "0.1.0"
description = "Framework open source para documentacao tecnica modular, rastreavel e auditavel."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "VerityDocs Contributors" }]
keywords = ["documentation", "traceability", "audit", "requirements", "prd", "spec"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/paulotks/VerityDocs"
Repository = "https://github.com/paulotks/VerityDocs"
Issues = "https://github.com/paulotks/VerityDocs/issues"
Changelog = "https://github.com/paulotks/VerityDocs/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest-cov>=5.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
veritydocs = "veritydocs.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/veritydocs"]

[tool.hatch.build.targets.wheel.force-include]
"src/veritydocs/py.typed" = "veritydocs/py.typed"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
